fix(goodeggs-deploy): actually support all Statsfiles
In https://github.com/goodeggs/travis-utils/pull/74, I tried to add missing support for
`Statsfile`s not written as `.js` files.

Unfortunately, this doesn't really work. Since we `require` the Statsfile, we need to be
configured with Babel, any hooks like `@babel/register`, and all that jazz, or it fails to
parse. This would be a big pain, since apps have different setups.

See example error at the bottom of this build:
https://ecru.goodeggs.com/projects/5a0e13c667ed811100ba2af3/builds/61b9db694e8c9600131ab858.

Instead, just `require` the built Statsfile. That's mildly confusing, since we don't
actually end up 'exercising' that while when we apply the Statsfile, but it seems OK.

I tried all this locally to verify.

Every app that uses this script except one builds its entire source to `build/`. I'm
opening a PR to fix the one exception.