fix: apply Statsfile on goodeggs-stats-cli version change
I just made a change to goodeggs-stats-cli and got it bumped in like 15 apps, only to
realize the Statsfile apply got skipped in all those apps. 😅

This is because we hash the export of the `Statsfile` and only apply it if has changed
since the last apply, so when we upgrade goodeggs-stats-cli, this hasn't changed, so we
skip the apply.

So, include the installed version in the hash.